--- main.tf
@@ -1,76 +1,0 @@
-terraform {
-  required_providers {
-    aws = {
-      source = "hashicorp/aws"
-    }
-  }
-}
-
-provider "aws" {
-  access_key = var.aws_access_key
-  secret_key = var.aws_secret_key
-  region     = var.aws_region
-}
-
-resource "tls_private_key" "example" {
-  algorithm = "RSA"
-  rsa_bits  = 4096
-}
-
-resource "aws_key_pair" "generated_key" {
-  key_name   = var.key_name
-  public_key = tls_private_key.example.public_key_openssh
-}
-
-resource "aws_security_group" "security_flugel_group" {
-  name        = var.security_group
-  description = "security group for flugel"
-
-  ingress {
-    from_port   = 8080
-    to_port     = 8080
-    protocol    = "tcp"
-    cidr_blocks = ["0.0.0.0/0"]
-  }
-
-  ingress {
-    from_port   = 22
-    to_port     = 22
-    protocol    = "tcp"
-    cidr_blocks = ["0.0.0.0/0"]
-  }
-
-  # outbound from server
-  egress {
-    from_port   = 0
-    to_port     = 65535
-    protocol    = "tcp"
-    cidr_blocks = ["0.0.0.0/0"]
-  }
-
-  tags = {
-    Name = var.security_group
-  }
-}
-
-resource "aws_instance" "flugelInstance" {
-  # count           = var.instance_count
-  ami             = var.ami_id
-  key_name        = var.key_name
-  instance_type   = var.instance_type
-  security_groups = [var.security_group]
-  tags = {
-    Name  = var.tag_name
-    Owner = var.tag_owner
-  }
-}
-resource "aws_s3_bucket" "flugel-s3-bucket" {
-  bucket_prefix = var.bucket_prefix
-  acl           = var.acl
-
-  versioning {
-    enabled = var.versioning
-  }
-
-  tags = var.tags
-}
