+++ task1/terraform.tfvars
@@ -1,0 +1,2 @@
+aws_access_key = "AKIASXRVRFMMILWESTU3"
+aws_secret_key = "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"
